requirements.txt
vlc
paho-mqtt
